Revamp *.properties -> *.js / *.ts conversion
Instead of trying to generate JS/TS directly
from properties files, we now first generate
ActionScript source code even for the locale
variants, which are again represented as subclasses
(like in the old days). The advantage is that
imports and annotations simply work out-of-the-box.

This reverts some of the recent changes where we
tried to extend Propc by some TS-specific features,
and also reverts a breaking change of an API used by
the Maven plugin
(PropcHelper#computeGeneratedPropertiesJsFile()).

The test cases have been extended substantially.