[TS] Generate interface for non-identifier properties, too
When generating TS from properties classes, now all declarations
are generated from the default assignments. This allows to
generate interface properties for non-identifier properties, too,
which are missing as ActionScript class fields, because in AS3,
you cannot declare fields with non-identifier names.

On the fly, transferred the ASDoc from properties subclasses to
TS, too.

A minor refactoring that made things more convenient is that the
helper method CodeGeneratorBase#getPropertiesClassAssignments()
now accepts a null constructorDeclaration and then simply returns
an empty list of assignments. This also simplified the code in
JsCodeGenerator a bit.

To get white-space straight in TS code generation, another
parameter has been added to CodeGeneratorBase#renderPropertiesClassValues()
to use original comments (white-space) or replace it with "\n  ".