Red test for subtle MXML import bug
Added a test that reproduces a subtle MXML import bug.
The trick is to import a class with the same name as the
current class in an <fx:Script> block. In that case, the
import incorrectly takes precedence over the declaration
of the class itself, so all unqualified usages of the
class name are incorrectly resolved to the imported class.
This can be observed even in JS output, where it apparently
did not matter that much, as the bug has gone undetected
for quite some years.