Fix MXML import bug
When using imports in an MXML <fx:Script> block, these
imports are not added to the compilation unit declaration
scope, but to the class declaration scope. Since scopes
are nested and all resolving takes place within the class,
this usually is no problem, but in case of name clashes,
it does. If a name clash is "distributed" over both scope,
it is not detected. Instead, class declaration scope "wins".
In this case, the inner import has even overridden the
declaration of the primary class declaration.