[TS] New configuration as3PackagePrefixToRemoveInTypeScript to shorten TS module paths
You may now set the Maven / compiler option as3PackagePrefixToRemoveInTypeScript
to the longest prefix that all AS3 packages in the current Maven module have
in common. This removes this redundant part from the path of the TS files
within the NPM package.
The desired prefix is stored in META-INF/pkg/package.json in each SWC. To access
this information, the compiler parses this JSON (ZipFileInputSource). To avoid
another 3rd-party dependency and because the JSON is generated by our own tool,
we parse it in a dirty line-by-line way :-(
TODO: Jangaroo can parse ActionScript object literals, so it can also parse JSON
as a special case. However, it is not so easy to apply the parser to a new file
type that *only* contains an object literal.

The main trick to implement the actual path shortening was to hand through the
prefix from Maven plugin to JoocConfiguration and from Jooc to the different
InputSource constructors and factory methods.
Then, IdeDeclaration#getTargetQualifiedNameStr() can access this information to
"shorten" the fully-qualified target name accordingly. When the prefix does *not*
match, currently, only an error message is printed to StdError, in order to debug
this "must not happen" case and still continue the whole conversion process.