[JS] Fix: Add extNamespace to class names of dynamic packages
For dynamic packages, the Packager creates a JSON file with a list
of all contained Ext classes. Here, the Ext class name is derived
from the relative path to the JS file. Since we now shorten that
path by the configured extNamespace, this prefix must be added
again when reverse-engineering the class name.

To achieve this, Packager needs the extNamespace as an additional
parameter, and since it is public API also used by the IDEA plugin,
we must be very careful when adding that parameter.
There is now a new interface Packager2 with a corresponding method
doPackage2(), which adds an extNamespace parameter.
PackagerImpl implements both the old and the new interface.
Thus, an updated IDEA plugin will still be able to use older
compilers by checking the created PackagerImpl instance for
instanceof Packager2 before calling the new API, else falling back
to the old API.