[JS/TS] Fix path ext namespace shortening for JS mode
Reverted IdeDeclaration#getTargetQualifiedNameStr() to its
old semantics to always return the Ext qualified name to target.
Instead of using a flag to cut off the ext namespace prefix
when required, introduced a new method
IdeDeclaration#getExtNamespaceRelativeTargetQualifiedNameStr()
(I know, that is a long name) that is only called when computing
relative the file path of a source file, be it for the output
file path (SingleFileCompilationUnitSinkFactory) or the TypeScript
ES5 module name (TypeScriptModuleResolver).
Maybe the remaining three usages could be centralized into a
utility method.
Note that the special check that a non-'required' compilation
unit need not be ext-namespace-shortened is gone, since the new
method should never be called with such a compilation unit,
because those CUs are [Native] and do not have individual output
files, but simply disappear in JS and go into index.d.ts in
TypeScript mode.