[JS] Fix bindable config property assignment
Since the compiler now guesses whether a variable is
of a config type (in TypeScript: Foo._), I tried to
use that information to leave out AS3.setBindable()
when the left-hand side of the dot-expression is known
to be a config. However, the check
  !leftHandSide.getType().isConfigType()
used the wrong expression, namely the whole left-hand
side (config.foo), not the dot-expression's left-hand
side (config).
While at it, simplified the code which seemed overly
complex due to some other cases which no longer exist.

Expected integration test code only had to be changed
because there was an unintended assignment (+=) in
TestComponentBase.as. We really should add tests for
AS3 assigment of (bindable) config object properties.