Use asConfig(obj).foo instead of obj.get/setConfig("foo")
This leads to better IDE support: Using the member key as a
string parameter works in theory, but all IDEs do not support
completion, but only show an error when a non-config is used.
Also, the new syntax saves us from rewriting assignments as
a special case, because it uses the same syntax on lhs and rhs.
Example: Assume foo is a [Bindable] config of obj, AS3 code

    var foo:String = obj.foo;
    obj.foo = "update";

now becomes

    import {asConfig} from "@jangaroo/joo/AS3";

    var foo:string = asConfig(obj).foo;
    asConfig(obj).foo = "update";

Like with AS3.cast(), it may happen that the new AS3.asConfig()
is imported, but never used. This is the case if some [Bindable]
property is accessed, but only through a variable of a Config type.
The code is then left a simple property access.