[TS] Pull mixin documentation from interface
Because the ActionScript mixin interface and implementation
class generate only one TypeScript class, documentation
at interface methods was lost during conversion.
Now, the TypeScript class uses the documentation of the
interface and its methods.
Note that this might overwrite documentation contained in
the AS mixin implementation class. For the whole mixin,
this happens without warning, but for mixin methods, a
warning is logged during conversion that mixin methods
must be documented in the interface, not the implementation
class. The warning looks like so:

 "Mixin method implementation has non-inheriting ASDoc.
  Please move such documentation to the mixin interface
  before TypeScript conversion."