[AST] Prevent [Bindable]s in MXML classes inheriting from joo.JavaScriptObject
It does not make sense to use [Bindable] configs in subclasses
of joo.JavaScriptObject, because inheriting from that class is
supposed to indicate that you do *not* want to use the Ext class
system, so you also do not want to use Ext's config system.
However, there are already MXML classes that inherit from
joo.JavaScriptObject. Here, MXML was just used as a more declarative
way to specify classes that only consist of properties.
The problem is that normally, all MXML <fx:Declarations> generate
[Bindable] properties.
Changed semantics:
* <fx:Declarations> in MXML subclasses of joo.JavaScriptObject now
  generate [ExtConfig] members rather than [Bindable] members
* Using explicit [Bindable] annotations in joo.JavaScriptObject
  subclasses, no matter whether they are MXML or AS, is an error