[TS] Fix: Allow extNamespace itself as targetName
Without this fix, trying to use 'Ext' of ext-as
raises an error that it does not start with the
configured namespace 'Ext', although it does.
The separating '.' must be handled accordingly.
The resulting "relative target qualified name str"
is then empty. To make the complete ES5 module
name look nice, TypeScriptModuleResolver checks
for such an empty relative module name and then
leaves out the trailing slash.