[TS] Fix: Really only suppress single space, not single white-space character
The failing ITs under Linux revealed that the check
for "one single space" was incorrect, as is was just a
check for "one single white-space character". Since
Linux newlines are one white-space character, but
Windows newlines are two, the check worked (mostly)
under Windows, but not under Linux.
When fixing this, even under Windows, some newlines
re-appeared, because they are generated and seem to
be a Linux newline even under Windows.