[Dep] Make undeclared used compile dependencies an error
Maven plugin has new option 'skipDependencyChecks' that also
reacts on system property 'enforcer.skip' to still be able
to build in case of dependency errors.
Internally, this sets the compiler's 'compilePath' to the
empty list, which in turn allows anything on the 'classPath'
to be used as a compile dependency. This also restores the
old behavior in case you do not specify a 'compilePath' on
the command line.