Enable check for unused declared dependencies
JoocConfiguration now lets isFindUnusedDependencies() return false
if dependency checks are skipped (compilePath.isEmpty()), and the
Maven plugin sets the flag only for production, not test code.
Thus, the unused dependency check can be triggered in Jooc exactly
when the config flag isFindUnusedDependencies() is set.

Tested by adding an unused dependency in a Studio module and seems
to work fine!