Fix multi-threaded execution of workspace converter
The problem was caused by the lock of the synchronized keyword
being based on a static variable and over the cause of the execution,
the object identity of that lock changed, making this mechanism useless.
This was because maven uses internally different classloaders during one
maven run. Because of this another object was used as lock that would
stay the same over the whole run.
Also do some minor code fixes and copy the static files instead of
writing them in the code.