[TS] Fix: Pull out variable declarations from constructor code wrapped into IEF
To trick TypeScript into not complaining about code using
'this' before the super constructor call, such code is
wrapped into an Immediately Evaluating Function (IEF).
However, if the code contains a variable declaration that
is used outside of the to-be-wrapped code, that variable
is not accessible there anymore.
To fix this, the to-be-wrapped directives are now scanned
for variable declarations. If a variable is declared _and_
used outside the to-be-wrapped code, its declaration is
not wrapped (kept directly inside the constructor body),
but only its initialization is moved inside the IEF.

Extended an integration test (SuperCallParameters) to
cover this special case.