[TS] Accept, but warn about wrong config constructor parameter type
Warn when the TypeScript converter encounters a class that seems
to use the Ext Config System, but does not declare a 'config'
constructor parameter of an allowed type: only the class itself,
one of its direct subclasses (usually only exactly one MXML class),
or the direct super class.
However, in contrast to the old behavior, the generated TypeScript
module now has a Config type (interface, namespace with '_') even
if the config type is invalid.