[TS] Suppress obsolete Exml/ConfigUtil.asString(<expr>) if expr is of type String
This intentionally changes the generated result of several test classes.
To also check the case where Exml.asString() is actually needed,
weakened the type of field 'myProperty' in test class AllElement.mxml
from String to *.
Note that AS3 API Exml.asString() is renamed to ConfigUtil.asString()
in TypeScript when using the actual ext-as API during conversion.