[TS] Improve converting __initialize__ return type to config type
Instead of rendering "._", we can mark the return expression type
as a config type (if possible) and it will then be rendered
correctly automatically.
Note that the check that the return type equals the class itself
is gone, we now assume that any return type of __initialize__ that
has a corresponding config type must be a config type.