[AST/TS] Report invalid __initialize__ return type as error
...also in JS mode. This is better than to get an error in
ActionScript code generated from MXML, where it is much harder
for the developer to pin down the original error.
To report the error independent of JS/TS mode made it necessary
to move isInitMethod to FunctionDeclaration, where this happens
during 'analyse'.
Added a test that the error is reported correctly.