[TS] Change Config Constructor back to Factory Function
Now that we evaluated Config Factory Functions as a
superior programming model, it makes more sense to
return to always creating Config objects by functions,
not pretending to instantiate a non-existing class.
The change to classes was justified by problems in IDEs
with declaration merging between a type and a function,
which seemed to be mitigated by using the usual combination
of type and object-with-constructor = class. But we could
not reproduce these problems anymore.

Thus, the namespaced '_' value is now a factory function
    (config?: _) => _
instead of an object-with-constructor = (pseudo) class
    { new(config?: _): _; }

Constructing Config objects happens by applying the
factory function
    SomeComponent._({ someConfig: someValue })
instead of instantiating the Config "class"
    new SomeComponent._({ someConfig: someValue })