[TS] Also convert named functions to arrow functions when name is unused
Just remember the variable declaration generated when a
function expression is named, then instead of checking
whether a name (ide) is given, check whether this variable
declaration is actually used.
This fixes a rare case where in MXML, a function expression
used a name, was not converted to an arrow function, used
$this instead of this, but $this was not declared + assigned.

Added extensive test for various function declarations
and expressions. This test also reveals that
    var this$ = this;
is still generated in TypeScript when it has become
obsolete because an arrow function has been used. But this
is not so easy to fix, as the "notification" that 'this'
has been used already happened and can't be taken back.