[AST=>TS] Remove unused var this$=this
To detect whether a function using the outer 'this' has
been converted to an arrow function, which is only possible
_after_ the analyse phase, we now register each FunctionExpr
at its surrounding method (FunctionDeclaration), so that
the isThisAliased() check of the method can be delegated
to all contained functions.
Because only TS mode supports arrow functions, a flag is
handed in to isThisAliased(). Alternatively, we could
generate arrow functions in JS mode, too, but then we
must remember to configure Sencha Cmd to allow them
(ES5 mode). Also, the generated JS code would look less
similar to the ActionScript code, but this similarity
is still an important goal of AS mode.