Remove everything EXML in Jangaroo 4.1
We should have done this a long time ago, but the
ext-as-api-generator for Ext AS 4 still had some
dependencies on EXML stuff. The API generator now
just lives on master-4, so we get delete all EXML
modules now.
Just one constant for the "exml:" Config URI prefix
had to be copied to AbstractCompilerMojo, and the
obsolete ManifestMojo has been deleted.