[AST=>TS] Simplified syntax for nested, typed object literals
In MXML, there may be <mixins> sub-elements and/or untyped
configs:
    <Foo someFooConfig="..." u:untypedConfig="...">
      <mixins>
        <SomeMixin someMixinConfig="..."/>
        ...
      </mixins>
    </Foo>

Previous TS syntax:
    new Foo._(_<Foo._ & SomeMixin._ & { untypedConfig }>({
      someFooConfig: ...,
      untypedConfig: ...,
      someMixinConfig: ...
    })

So the previous syntax focussed on declaring a single, flat
object literal, typed with an ad-hoc type composed of the
participating types.
We came up with a new syntax that is easier to read and
write and closer to the original MXML syntax.
The 'mixins' sub-elements are represented as sub-objects,
using the spread operator to "flatten" it into the main
config object literal.
The same trick is used for untyped configs, only no type
assertion is used.
With that the example above becomes
    new Foo._({
      someFooConfig: ...,
      ...{
        untypedConfig: ...
      },
      ..._<SomeMixin._>({
        someMixinConfig: ...
      }
    })
It is now much clearer, even without an IDE, where each
config comes from. 'Foo._' does not have to be repeated.
The names of untyped properties do not have to be repeated.
The complex type operator '&' is not used anymore.