Consider [Return] annotation in function signature
Although the annotation was mainly meant for TypeScript migration,
it is now also considered when determining the function signature.
This is important for change c1b4b2eb, where functions with return
type void lead to simpler arrow functions, because there are
methods with return type void, but a [Return("this")] annotation,
leading to a non-void return value. For such methods, "handing
through" the void value is not applicable, so the arrow function
must keep using a block.