Fix MXML attribute value conversion
Computing a typed Java value from an MXML attribute value, then
turning it back into a string was a bad idea. For example, it
resulted in the MXML attribute value "1" assigned to a property
of type Number to be rendered as "1.0". Also, all kinds of values
were accepted as Booleans and _everything_ not "true" (case
insensitive) was silently converted to "false".
Now, we only type-check the value, but keep it as-is. This may
(later) result in an ActionScript type error, but if not, the
original representation of the value is kept.
For example, when using foo="1" in MXML, the generated AS code
and thus the resulting JS or TS code will now all use
    foo: 1,
not
    foo: 1.0,
or any other "normalized" value.