Fix missing const this$=this for nested functions
The logic when to render 'const this$=this;' did not
work for nested function expressions.
Instead of registering all function expressions at
their containing method, they now register 'this'
usage at their direct parent function expression and
propagate this usage there.
When rendering the method, it delegates the decision
whether to "alias 'this'" to its direct child functions,
which again search their direct children and so on.

Added a test case for nested function expressions,
where the out function cannot be rewritten to an
arrow function, but the inner one can. This used to
result in a missing 'const this$=this;', but no
longer does so.