Prevent trailing comma when <exml:mixins> contains empty object
If a Spread of an empty object is "inlined" in JS mode, a
trailing comma was generated, which confused Sencha Cmd.
The root cause for the empty object, however, is that empty
MXML attributes of Object-typed properties generate no
ObjectFields at all, because createValueExprFromTextSymbol()
returns null. Looking at some MXML code, it seems that this
used to be different, but I am not sure what default value
was used - empty string, empty Object, null?
Must investigate...