Fix 'this' aliasing with usages inside a loop
Loops have their own scope to locally define labels, so
the scope handed over to FunctionExpr#notifyThisUsed() was
not as expected the function's body scope.
As a FunctionExpr already knows its body scope, the parameter
can be removed and bodyScope can be used directly.

Added another test to also check this setup:
Implicit 'this' usage inside a loop inside a function.