[TS] Fix: MXML element with only untyped attributes
Example:
    <Foo u:untyped="foo"/>

The generated TypeScript code
    new Foo({...{ untyped: "foo"}})
used to lead to a type error. The trick to avoid the type
error for untyped properties only works when there is at
least one typed property.
So now we generate
    new Foo(<Foo._>{ untyped: "foo" })
which is allowed because of the more lax rules for type
assertions on object literals.