[JS+TS] Fix: field initializer function using outer this
Previously, the outer 'this' access was aliased (JS: _this,
TS: this$), but the alias was never assigned.
It is a bit tricky to get hold of the outer 'this' and declare
the alias, and there a slightly different solutions for JS
and TS.
For JS, there is already a dedicated initializer function,
in which outer 'this' is still accessible and _this can be
defined.
For TS, we use a standard ECMAScript field initializer with
the function as its direct value, so a local variable would
cause some syntactic overhead. Instead, we define an
immediately evaluating arrow function and hand the outer
'this' into its parameter this$.