[TS] Change: treat [Lazy] const like var => use '_' property
To be able to initialize a package-scope const in a
truly lazy fashion, [Lazy] const-s are now, like var-s,
converted to an object with a '_' property. This '_' is
an accessor that calls the initializer function on first
read access. Lazy var-s also have a set accessor. The actual
implementation will be added to jangaroo runtime as utility
functions 'lazyConst' and 'lazyVar'.
Thus, the TypeScript definition of a [Lazy] const changes from
  const foo: <type> = <init>;
to
  const foo: { _: <type> } = lazyConst(() => <init>);

Access changes from 'foo' to 'foo._'.
The access is no longer rewritten by the Babel plugin, so that
is actually happens at runtime. This makes __lazyFactory__
obsolete; only __factory__ will be used in Jangaroo NPM.