[TS] Fix: Also rewrite {...expr} in type assertion function argument
When the "type-check object literal" function is called with an
object literal that only contains a spread, the same logic as for
type casts has to be applied: the inner expression must be unwrapped
and a type assertion must be added.
To allow reusing the logic from visitApplyExprArgument(), refactored
it to renderSingleSpreadValue(), which now returns a flag whether
it has rendered a simplified expression, so that else, different
output can be generated.
Also simplified rendering the type to check: when it is already marked
as a config type in ApplyExpr#analyze(), we can simply use
getTypeScriptTypeForActionScriptType() and no longer need to retrieve
the type declaration and render '._' explicitly.