[TS] Special treatment for Config objects of non-Ext classes
When a non-Ext class (a class not explicitly inheriting
from Ext.Base) without a Config type is used to type-cast
an empty object literal, in TypeScript, we now call
the no-arg constructor instead of creating a Config object
with a target class marker ('xclass' etc.):
    new Foo()
instead of
    Config(Foo)

When a non-Ext class with a Config type is used in a type
cast, instead of adding a target class marker to the Config
object, we now just call the Config type assertion function:
    Config<Foo>({ ... })
instead of
    Config(Foo, { ... })

Even for Ext classes, using a config constructor parameter
with an interface type no longer derives a Config type.

These changes together should make non-Ext class Configs
work in "real" ECMAScript.