[TS] Fix: import path uses letter case of existing directory
When computing the relative path to use to import a TypeScript
module, CompilerUtils.getRelativePath() is used, which again
uses File#getCanonicalPath(). If in the source directory tree,
a directory corresponding to a namespace part already exists,
under Windows, the letter case of that directory is used, not
the desired letter case of the target namespace. In the tests,
this happens for Ext -> ext.
To fix this, we now use Path#relativize() rather than
CompilerUtils.getRelativePath(), which does not care for
existing directories at all.