Use Config instead of cast when parameter has Config type
Even when the parameter of a type cast of a class with a
Config type is _not_ an object literal, if it is an argument
to a function call (ApplyExpr), consider the corresponding
parameter type. If the parameter type is a Config type, the
type cast result must be a Config object, so the type cast
better use 'Config', not 'cast'.
This requires to add 'Config' during analyse phase in case
a type cast may be rendered as 'Config'. While at it, also
suppressed some 'cast' imports that become 'Config', anyway.
This caused the only change to the expected test code compile
results.