Fix: let 'declare Config' always reference own Config interface
The generated code tries to reference the non-exported interface
of the subclass directly by its module-local name, which is bound
to fail.
Referencing the Config interface of the subclass through the Config
utility type is also not possible here, because this would cause
a circular definition of the subclass' Config interface, which
extends the Config interface of the current class.
Thus, 'declare Config' must always reference the locally defined,
"own" Config interface.