Fix: missing 'Config' import for 'config' identifier
When a Config type is guessed from an identifier (mostly
parameter) is named 'config', this happens late, during
render phase, so it is too late to add an import of the
'Config' utility type. This fix (or rather workaround)
adds the import during analyze phase, whenever an
identifier is named 'config'. While this may add the
import too often, an unused import is later eliminated
by eslint-typescript.