Fix: Add ._ when accessing var or lazy singleton via "reflection"
By recognizing the API call joo.getOrCreatePackage("..."), access to
a package-scope var or lazy const is now recognized although it is
done by "reflection", i.e. dynamic. To even catch usages of the
renamed singleton 'SESSION' (renamed to 'session'), a heuristic
tries to find a CU in the same package, but with the local name in
upper case, and if found, checks its [Rename] annotation.
Since such reflection access is primarily used in JooUnit tests to
mock singletons, this fixes the remaining red JooUnit tests in
Studio Client.