[TS] Generate static class code blocks (ECMAScript proposal)
ActionScript static class code was formerly represented
in TypeScript by a static private member with a
self-evaluating initializer.

Since version 4.4, TypeScript supports the ECMAScript
proposal for static class code blocks.
We now use these to represent AS3 static class code.
This simplifies the TypeScript output a bit, but we
must move the static code block to the end of the class,
as forward references to static members are not allowed
in ECMAScript (but they were allowed in ActionScript).