[TS] Use '._' even for non-lazy const-s
Rationale is that even package-scope variables
declared as 'const' are written sometimes, via
a "trick" using
  joo.getOrCreatePackage("<package>").<ide> = ...
This workaround code can only be rewritten to
ECMAScript/TypeScript if there is a property
to assign (here: '_'), because you cannot
reassign a module export.
The last remaining case where there is *no*
'_' in TypeScript is when using [Native]
package-scope const-s or var-s.