[TS] Allow writing package-scope const via getOrCreatePackage()
Rationale is that even package-scope variables
declared as 'const' are written sometimes, via
a "trick" using
  joo.getOrCreatePackage("<package>").<ide> = ...;
or
  joo.getOrCreatePackage("<package>")['<ide>'] = ...;

This workaround code can be rewritten to TypeScript
by casting to the writable version of the original
type:
  import <ide> from "...";
  (<ide> as { _: typeof <ide> })._ = ...;
The type cast ensures that even though you assign
to a read-only property, you at least use a value
of the correct type.
Casting to a non-related type would be an error
detected by TypeScript, too.

Even deleting the current value works, if we
make the '_' property optional:
    delete joo.getOrCreatePackage("<package>").<ide>;
now becomes
    import <ide> from "...";
    delete (<ide> as { _?: any })._;