[TS] Replace Ext.apply() by Object.assign()
Replace usages of Ext.apply(x, y[, z]) by Object.assign(x[, z], y).
Semantics are slightly different, but only for a corner case
we are not interested in, namely that the first argument is _not_
an object (primitive or null or undefined).
The import of Ext is not removed by the compiler, but eventually
by the conversion process by invoking eslint.