[TS] Rewrite init function to top-level code
With a working lazy initialization, it is no longer
possible to execute certain code on load.
When a Maven module / npm package wants to do so,
it must now define an init function in its extNamespace
top-level package. So for example, if extNamespace
is 'com.acme', you would define the init function
in src/main/joo/com/acme/init.as, so that the result
will end up in target/package/.../init.ts.
While the init code is given as a function in AS
(because there is no such thing as top-level module
code in AS), it is unwrapped to top-level module code
in TS.
The bundling will then take care of adding the init.ts
module in a way that it is executed when the whole
package code is loaded.