[TS] Fix: cut off extNamespace of module referenced by pseudo-singletons
The module to import instead of the pseudo-singleton is derived
from its [Native(..., require="<module-to-import>")] annotation.
Of course, this value is an "absolute" namespace, so the extNamespace
prefix has to be removed before using it as npm path.
This bug did not occur in the integration test, since the Ext AS
API is "faked" there, not in a separate Maven module with its own
extNamespace.
Fixing the path led to confusion between the pseudo singleton and
the compilation unit of its type, resulting in usages of the pseudo
singleton no longer being replaced by <type>.getInstance().
This is fixed by resolving the type compilation unit when computing
local names (for name clashes) and in the second phase when creating
the import name mapping.