[TS] Different logic for Ext events and Flex Events
There is some confusion with how we represent Ext
events in ActionScript to make them Flex compatible.
To define Flex events with the [Event] annotation,
the Ext event name gets prefixed with "on" to avoid
name clashes with Config properties, because in
MXML, you can then add 'onMyEvent="<event handler
function body>"' as an attribute like you would
for the property 'onMyEvent'.
Besides the 'name', an [Event] annotation must
specify a 'type', which points to a class
representing the _sole_ event callback parameter.
Since Ext works with complex event callback
signatures, Jangaroo's MXML compiler generates
adapter code based on the declared (readonly)
fields of the referenced event type class.
The selection and order is determined by a
special static const __PARAMETER_SEQUENCE__,
but since its value is not avaible in the API
stubs, the declarations of these fields should
be in the exact same order. All such "fake"
event classes must inherit directly from
net.jangaroo.ext.FlExtEvent.
MXML also requires an @eventType annotation in
the ASDoc of the [Event] annotation, pointing
to a constant in the event type class defining
the event name (redundantly).

The AS->TS Converter now acts differently on
[Event] annotations with a type pointing to
a net.jangaroo.ext.FlExtEvent subclass than
when there is no type or the type is not a
FlExtEvent. If it is a FlExtEvent, the mapping
described above is reversed, so 'onMyEvent'
becomes the Ext event 'myevent'. The event
callback signature is derived from the event
type class as described above.
Otherwise, the event name is taken as-is and
the event signature will actually just take
the single parameter of the event type.

Had to fix some compiler integration test code
to actually match the pattern described above
and fix a bug in reverse-engineering the event
name from the constant name.