[TS] Fix: Find [Event]s in mixin interfaces
Since events are part of the API of a mixin,
they should be declared in the mixin interface,
not the mixin implementation class. So far,
[Event] annotations in the mixin interface
were ignored.
To actually be MXML compatible, you would
have to copy all [Event] annotation from mixin
interfaces to all classes that mix it in.
This again shows that ActionScript is unable
to really cope with mixins.