[TS] Create Config interface for listeners: MyEvents
If there is no config parameter in the constructor, usually,
no Config interface is created. But if custom Ext Events
are declared, we need at least the 'listeners' Config
property that refers to the custom ...Events interface.
So even if there are only Events, no Configs, an own
Config interface must be declared (and also referred to by
'declare Config: ...').

On the fly, cleaned up outdated 'initialConfig' query that
was only needed back when the Config type was retrieved
from the type of 'initialConfig', not 'Config'.