[TS] Fix: add 'override' to REST_RESOURCE_URI_TEMPLATE when appropriate
The AS3 annotation [RestResource] generates a TS constant
REST_RESOURCE_URI_TEMPLATE. In TS, static member that redefine
a static member from a superclass need an 'override' modifier.
Of course, this was not detected for this very special case of
a generated constant, so we must check for the annotation in any
superclass explicitly to add an 'override' modifier when
appropriate.