Convert square-bracket property access to dot if possible
In ActionScript, square-bracket property access leads to
untyped usage of a property (at least in IDEA and Jangaroo,
not so sure about "real" ActionScript runtime).
In TypeScript, however, square brackets do not suffice:
you must add an 'as unknown' type assertion to disable
type checks completely.
So when the compiler finds an untyped access using a quoted
identifier like 'obj["field"]' and can resolve a typed
field 'field' from the type of 'obj', it rewrites this
to '(obj as unknown)["field"]'.
But there are several situations in which this untyped
access is not or no longer necessary. Here are the
different situations:
1. Write a read-only field - this must use untyped
   access in TypeScript, too, unless the code is in
   the constructor of the class defining the field.
2. Access a [Bindable] field, disabling rewriting to
   get/set methods / 'asConfig(obj).field'.
2. Use the field with an incompatible type - this
   should rather be expressed by type casts / 'as'
   expressions.
So now, we rewrite such untyped access to
    obj.field
unless it is a left-hand side _and_ the field is read-only,
unless this code is contained in a constructor.
We log a warning about rewriting the code, because it
is recommended to change the original ActionScript code.
In the worst case, we produce another TypeScript type
error, but since there are type errors, anyway, and they
can be ignored for the time being, this is not a
disadvantage of the new rewriting. To the contrary, if
you do not fix such untyped usages before conversion,
you should do it in TypeScript, and the type error
reminds you of that.