Improve generated jangaroo.config.js pretty-printing
Instead of using search/replace in a JSON string, we now customize
the pretty printer to generate trailing commas in the first place.
Unfortunately, Jackson's DefaultPrettyPrinter is not sufficently
flexible to use a similar approach for not quoting keys that have
identifier syntax. DefaultPrettyPrinter knows a flag
com.fasterxml.jackson.core.json.JsonWriteFeature.QUOTE_FIELD_NAMES
but that flag is deprecated and when switched off, _never_
quotes the keys, even if they do _not_ use identifier syntax.
There even is a TODO in Jackson's code to enforce quoting for
non-identifiers.
Also, DefaultPrettyPrinter provides no hook to override key
rendering (with or without quotes, depending on the key), so
unless you implement a custom serializer almost from scratch,
there is no way to un-quote identifier keys.
Thus, we let the PrettyPrinter render quoted keys and un-quote
identifier keys afterwards using a RegExp as before.

On the fly, simplified PackageJsonPrettyPrinter a bit:
There is a protected field _objectFieldValueSeparatorWithSpaces
that can be overwritten once to customize that we want a
space only after the colon.

Added a unit test for jangaroo.config.js pretty-printing.