Fix type of NewExpr without parenthesis
When using 'new Foo' instead of 'new Foo()', the NewExpr
type was incorrectly resolved to be Class<Foo>, not Foo
(the instance type).
While rare, it is allowed, even in TypeScript, to call a
constructor implicitly, i.e. leave out the parenthesis.
Modified an integration test to test this case: If 'new'
is used without parenthesis, the type on the left-hand
side of the assignment is only left out if the NewExpr
type is resolved correctly.