Fix: Add complementary accessor when overriding only get or set
In ActionScript, if you override an accessor, but not its complementary accessor (i.e. override
set, but not get), the complementary accessor is inherited. In ECMAScript, it is not, so we must
generate TypeScript code to also override it and simply delegate to the super property.
Added integration test Jooc.testOverrideOneAccessorOnly():
ConfigSubclass overrides only the get accessor of 'defaultType' and only the set accessor of
'title' (a [Bindable] property).